dump of found form component
Wrote tests against it and clean up a little. Not all test pass; specifically it does not validate all fields on submission. This is problematic if a field conditionally validates based on another field that is not filled out at the time the validating field changes/blurs. See failing tests for example.